Split XML tests into gui/headless
